chore(main): extract settings-runtime-effects from main.ts (arch R5)
Pure move of main.ts's settings runtime-effects cluster into
settings-runtime-effects.ts, exporting createSettingsRuntimeEffects(deps)
→ { normalizeSettingsPatch, applySettingsRuntimeEffects,
handleExternalSettingsChange } (syncDefaultPermissionModeToSessions stays
internal). Follows the R4 tool-assembly / tool-artifact-persistence DI
precedent: settingsStore / botRegistry / openGateway / keepSystemAwake /
runtime / safeSendToRenderer / emitSessionsChanged are injected, bodies
byte-identical. keep-awake runtime-effect (#1207) rides
applySettingsRuntimeEffects unchanged.

Contract mechanics: add settings-runtime-effects.ts to the
main-process-contract-source-helpers aggregator; allowlist the moved
[config-watcher] console.error at its new path in check-console.mjs. No
source-pin test targeted these functions in main.ts directly.
