+++ apps/desktop/src/main/__tests__/main-process-contract-source-helpers.ts
@@ -30,6 +30,7 @@ export const MAIN_PROCESS_SOURCE_REPO_PATHS: readonly string[] = [
   'apps/desktop/src/main/project-root-controller.ts',
   'apps/desktop/src/main/session-entry-ipc-main.ts',
   'apps/desktop/src/main/session-branch.ts',
+  'apps/desktop/src/main/settings-runtime-effects.ts',
   'apps/desktop/src/main/sessions-ipc-main.ts',
   'apps/desktop/src/main/settings-ipc-main.ts',
   'apps/desktop/src/main/subscription-model-fetch.ts',
